requirements.txt
svdtools~=0.1.0
importlib_resources ; python_version < '3.7'
loguru~=0.5.0
PyYAML~=5.3.1